EcoPatrol/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
asgiref==3.2.10
pytz==2020.1
sqlparse==0.4.1
drf-extensions
drf-spectacular
pillow==8.0.1
psycopg2-binary
django-filters